+++ opal/caregivers/models.py
@@ -27,6 +27,7 @@ class CaregiverProfile(models.Model):
     legacy_id = models.PositiveIntegerField(
         verbose_name=_('Legacy ID'),
         validators=[MinValueValidator(1)],
+        unique=True,
         null=True,
         blank=True,
     )