+++ opal/core/api_urls.py
@@ -9,7 +9,7 @@ from django.urls.conf import include
 
 from rest_framework.routers import DefaultRouter, SimpleRouter
 
-from opal.caregivers.api.views import GetCaregiverPatientsList, GetRegistrationEncryptionInfoView
+from opal.caregivers.api import views as caregivers_views
 from opal.caregivers.api.viewsets import SecurityAnswerViewSet, SecurityQuestionViewSet
 from opal.core.api import views as core_views
 from opal.hospital_settings.api import viewsets as settings_views
@@ -40,7 +40,7 @@ urlpatterns = [
     path('app/chart/<int:legacy_id>/', AppChartView.as_view(), name='app-chart'),
     path('app/home/', AppHomeView.as_view(), name='app-home'),
     path('auth/', include('dj_rest_auth.urls')),
-    path('caregivers/patients/', GetCaregiverPatientsList.as_view(), name='caregivers-patient-list'),
+    path('caregivers/patients/', caregivers_views.GetCaregiverPatientsList.as_view(), name='caregivers-patient-list'),
     path('languages/', core_views.LanguagesView.as_view(), name='languages'),
     path(
         'patients/legacy/<int:legacy_id>/check_permissions/',
@@ -52,10 +52,24 @@ urlpatterns = [
         CaregiverRelationshipView.as_view(),
         name='caregivers-list',
     ),
-    path('registration/by-hash/<str:hash>/', GetRegistrationEncryptionInfoView.as_view(), name='registration-by-hash'),
+    path(
+        'registration/by-hash/<str:hash>/',
+        caregivers_views.GetRegistrationEncryptionInfoView.as_view(),
+        name='registration-by-hash',
+    ),
     path('questionnaires/reviewed/', QuestionnairesReportView.as_view(), name='questionnaires-reviewed'),
     path('app/general/', AppGeneralView.as_view(), name='app-general'),
     path('registration/<str:code>/', RetrieveRegistrationDetailsView.as_view(), name='registration-code'),
+    path(
+        'registration/<str:code>/verify-email/',
+        caregivers_views.VerifyEmailView.as_view(),
+        name='verify-email',
+    ),
+    path(
+        'registration/<str:code>/verify-email-code/',
+        caregivers_views.VerifyEmailCodeView.as_view(),
+        name='verify-email-code',
+    ),
     path(
         'caregivers/<uuid:uuid>/security-questions/',
         SecurityAnswerViewSet.as_view({'get': 'list'}),
@@ -80,6 +94,11 @@ urlpatterns = [
         'institutions/<int:pk>/terms-of-use/',
         settings_views.InstitutionViewSet.as_view({'get': 'retrieve_terms_of_use'}),
         name='institutions-terms-of-use',
+    ),
+    path(
+        'caregivers/devices/<str:device_id>/',
+        caregivers_views.UpdateDeviceView.as_view(),
+        name='devices-update-or-create',
     ),
     path('', include(router.urls)),
 ]