+++ opal/legacy/factories.py
@@ -123,6 +123,13 @@ class LegacyQuestionnaireFactory(DjangoModelFactory):
     completedflag = 0
 
 
+class LegacyPostcontrolFactory(DjangoModelFactory):
+    """Post Controle factory for announcement from the legacy database."""
+
+    class Meta:
+        model = models.LegacyPostcontrol
+
+
 class LegacyAnnouncementFactory(DjangoModelFactory):
     """Announcement factory from the legacy database."""
 
@@ -130,6 +137,7 @@ class LegacyAnnouncementFactory(DjangoModelFactory):
         model = models.LegacyAnnouncement
 
     patientsernum = SubFactory(LegacyPatientFactory)
+    postcontrolsernum = SubFactory(LegacyPostcontrolFactory)
     readstatus = 0
 
 
@@ -144,7 +152,6 @@ class LegacySecurityQuestionFactory(DjangoModelFactory):
     questiontext_fr = 'Quel est le nom de votre premier animal de compagnie?'
     creationdate = timezone.make_aware(datetime(2022, 9, 27))
     lastupdated = timezone.make_aware(datetime(2022, 9, 27))
-
     active = 1